% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Splits.R
\name{splits}
\alias{splits}
\title{Splits a Dataframe}
\usage{
splits(data, ratio)
}
\arguments{
\item{data}{A dataframe}

\item{ratio}{Defined as a number, x, between (0,1) where x+y = 1 and x:y is the ratio of training:testing partitions Refer to sample.split documentation for more details.}
}
\value{
splits of given dataframe.
}
\description{
Splits a Dataframe
}
