% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseByCol.R
\name{summariseByCol}
\alias{summariseByCol}
\title{Create summary table for a dataframe}
\usage{
summariseByCol(df, col, target)
}
\arguments{
\item{df}{A dataframe}

\item{col}{Defined as the column name to group by}

\item{target}{Defined as the column to count}
}
\value{
summary table of dataframe which is group by a column, then counted by the target coluumn
}
\description{
Create summary table for a dataframe
}
